requirements.txt
python-dotenv >= 1.0.0
openai
datetime
# psycopg2 # For production
psycopg2-binary # For Development
colorama